Implement user certificate security #82
