+++ back-end/src/App.ts
@@ -1,6 +1,5 @@
 import express from 'express';
 import https from 'https';
-import tls from 'tls';
 import { Database } from './model/Database'
 import { router } from './routes/Router';
 import * as PassportConfig from './controllers/PassportConfig';
@@ -23,30 +22,14 @@ let db = new Database()
 db.connect("mongodb://localhost:27017/slowvid");
 
 app.use('/api', router);
+
 app.get('/', function (req, res) {
   res.send('Hello World!');
 });
 
-app.get('/authenticate', (req, res) => {
-	const cert = (req.socket as tls.TLSSocket).getPeerCertificate();
-
-	if (Object.keys(cert).length > 0 && (req.socket as tls.TLSSocket).authorized) {
-		res.send(`Hello ${cert.subject.CN}, your certificate was issued by ${cert.issuer.CN}!`);
-
-	} else if (cert.subject) {
-		res.status(403)
-			 .send(`Sorry ${cert.subject.CN}, certificates from ${cert.issuer.CN} are not welcome here.`);
-
-	} else {
-		res.status(401)
-		   .send(`Sorry, but you need to provide a client certificate to continue.`);
-	}
-});
-
-
 // listens on https
-https.createServer(Conf.httpsOptions, app).listen(4433, () => {
-  console.log("Listening at https://localhost:4433");
+https.createServer(Conf.httpsOptions, app).listen(4000, () => {
+  console.log("Listening at https://localhost:4000");
 });
 
 app.listen(3000, function () {
